requirements.txt
streamlit
python-dotenv
ollama
weaviate-client
pypdf
google-generativeai